[FIX] monitor step 계산 로직 변경

+++ module-monitor/src/main/java/com/xiilab/modulemonitor/repository/PrometheusRepository.java
@@ -5,5 +5,5 @@ public interface PrometheusRepository {
 
 	String getHistoryMetricByQuery(String promql, String startDate, String endDate);
 
-	String getHistoryMetricByQuery(String promql, String startDate, String endDate, int step);
+	String getHistoryMetricByQuery(String promql, String startDate, String endDate, long step);
 }
