+++ module-monitor/src/main/java/com/xiilab/modulemonitor/service/PrometheusServiceImpl.java
@@ -34,7 +34,7 @@ public class PrometheusServiceImpl implements PrometheusService{
 
 	@Override
 	public List<ResponseDTO.HistoryDTO> getHistoryMetric(RequestDTO requestDTO) {
-		int step = DataConverterUtil.getStep(requestDTO.startDate(), requestDTO.endDate());
+		long step = DataConverterUtil.getStep(requestDTO.startDate(), requestDTO.endDate());
 		// 검색시간 UnixTime로 변환
 		String startDate = DataConverterUtil.toUnixTime(requestDTO.startDate());
 		String endDate = DataConverterUtil.toUnixTime(requestDTO.endDate());
